change dir error fix

+++ src/migrate.sh
@@ -43,6 +43,8 @@ BACKUP_DIR=`dirname $INPUT`
 
 #safely get full path of backup directory ( this method is use to handle directories like ~/ )
 cd $BACKUP_DIR
+[ $? -ne 0 ] && { echo "[ ERROR ] [ ${LINENO} ]"; exit 1; }
+
 BACKUP_DIR=`pwd`
 
 STRIP_EXTN1="${BACKUP_FILE%.*}"
@@ -85,6 +87,7 @@ echo
 
 # Enter backup directory
 cd $BACKUP_DIR
+[ $? -ne 0 ] && { echo "[ ERROR ] [ ${LINENO} ]"; exit 1; }
 
 # Remove Old Restore Folder ( if any )
 rm -rf $RESTORE_DIR
@@ -125,6 +128,7 @@ echo
 
 #safely get full path of website root ( this method is use to handle directories like ~/ )
 cd $NEW_WEBSITE_ROOT_DIR
+[ $? -ne 0 ] && { echo "[ ERROR ] [ ${LINENO} ]"; exit 1; }
 NEW_WEBSITE_ROOT_DIR=`pwd`
